Address comment: inline particles implementation and manifest.
