--- src/runtime/test/artifacts/test-hosted-particles.manifest
@@ -1,25 +1,0 @@
-// Copyright (c) 2019 Google Inc. All rights reserved.
-// This code may only be used under the BSD style license found at
-// http://polymer.github.io/LICENSE.txt
-// Code distributed by Google as part of this project is also
-// subject to an additional IP rights grant found at
-// http://polymer.github.io/PATENTS.txt
-
-schema Foo
-  Text value
-
-particle UpdatingParticle in 'updating-particle.js'
-  out Foo innerFoo
-
-interface TestInterface
-  out Foo *
-
-particle MonitoringParticle in 'monitoring-particle.js'
-  host TestInterface hostedParticle
-  out Foo foo
-
-recipe
-  use as h0
-  MonitoringParticle
-    foo = h0
-    hostedParticle = UpdatingParticle
